유저, 메시지 상태 분리 후 컴포넌트 반영
