+++ src/App.tsx
@@ -2,9 +2,9 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import HomePage from './components/homePage/homePage';
 import ProfilePage from './components/profilePage/ProfilePage';
-import ChattingPage from './components/chattingPage/ChattingPage';
+import ChattingListPage from './components/chattingListPage/ChattingListPage';
 import SettingPage from './components/settingPage/settingPage';
-import MessangerPage from './components/messangerPage/MessangerPage';
+import ChattingPage from './components/chattingPage/ChattingPage';
 
 function App() {
   return (
@@ -12,10 +12,10 @@ function App() {
       <Routes>
         <Route path="/" element={<HomePage />}>
           <Route path="profile" element={<ProfilePage />} />
-          <Route path="messageRoomList" element={<ChattingPage />} />
+          <Route path="chattingList" element={<ChattingListPage />} />
           <Route path="setting" element={<SettingPage />} />
         </Route>
-        <Route path="messageRoom/:username" element={<MessangerPage />} />
+        <Route path="chattingRoom/:userId" element={<ChattingPage />} />
       </Routes>
     </BrowserRouter>
   );
