+++ src/components/chattingListPage/ChattingListPage.styled.tsx
@@ -1,0 +1,1 @@
+import styled from 'styled-components';
